third_party/WebKit/LayoutTests/http/tests/loading/sync-xhr-priority-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
http://127.0.0.1:8000/resources/test.mp4 has priority VeryHigh
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
PASS successfullyParsed is true

TEST COMPLETE
